backend/pyproject.toml
[project]
name = "fintrid-backend"
version = "0.1.0"
description = "Backend API for Fintrid financial application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "python-multipart>=0.0.12",
    "altair>=5.5.0",
    "attrs>=25.4.0",
    "blinker>=1.9.0",
    "cachetools>=6.2.1",
    "certifi>=2025.10.5",
    "charset-normalizer>=3.4.4",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "gitdb>=4.0.12",
    "gitpython>=3.1.45",
    "idna>=3.11",
    "jinja2>=3.1.6",
    "jsonschema>=4.25.1",
    "jsonschema-specifications>=2025.9.1",
    "markupsafe>=3.0.3",
    "narwhals>=2.10.2",
    "numpy>=1.26.4",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "pillow>=10.4.0",
    "protobuf>=6.33.0",
    "pyarrow>=21.0.0",
    "pydeck>=0.9.1",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "referencing>=0.37.0",
    "requests>=2.32.5",
    "rpds-py>=0.28.0",
    "six>=1.17.0",
    "smmap>=5.0.2",
    "streamlit>=1.51.0",
    "tenacity>=8.5.0",
    "toml>=0.10.2",
    "tornado>=6.5.2",
    "typing-extensions>=4.15.0",
    "tzdata>=2025.2",
    "urllib3>=1.26.20",
    "watchdog>=6.0.0",
    "landingai>=0.3.49",
    "pypdf2>=3.0.1",
    "langchain>=1.0.5",
    "plotly>=5.18.0",
    "sqlalchemy>=2.0.44",
    "rapidfuzz>=3.14.3",
    "pymupdf>=1.26.6",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.2.1",
    "landingai-ade>=0.20.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/fintrid_backend"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[project.scripts]
fintrid-api = "fintrid_backend.main:main"
fintrid-ui = "streamlit:run_streamlit"

[project.entry-points."streamlit.run"]
fintrid_backend = "fintrid_backend.streamlit_app:main"
